requirements.txt
azure_storage==0.37.0
beautifulsoup4==4.12.2
bert_extractive_summarizer==0.10.1
Flask==2.3.2
google_api_python_client==1.7.2
google_auth_oauthlib==1.0.0
imgkit==1.2.3
moviepy==1.0.3
newspaper3k==0.2.8
openai==0.27.4
pandas==2.0.1
Pillow==9.5.0
Requests==2.31.0
thefuzz==0.19.0
tqdm==4.65.0
transformers==4.28.1
